test(cketh): move the anvil client into test_utils and let the harness own it
Relocate the anvil node client and its ABI/solc helpers from the
deposit_from_cex test into ic_cketh_test_utils::live_scan, and have
CkErc20LiveScanSetup start and own the anvil node itself: new_live()
takes no URL, configures the EVM RPC canister to route to its node, and
exposes credit_deposit to seed the supported tokens for a deposit
address.

The standalone batcher tests now use the shared Anvil client from
test_utils, and the end-to-end test reduces to a thin script.
