+++ rs/ethereum/cketh/test_utils/BUILD.bazel
@@ -39,6 +39,7 @@ rust_library(
         "@crate_index//:hex",
         "@crate_index//:ic-cdk",
         "@crate_index//:num-traits",
+        "@crate_index//:reqwest",
         "@crate_index//:serde",
         "@crate_index//:serde_bytes",
         "@crate_index//:serde_json",
